requirements.txt
streamlit
ultralytics
opencv-python
pillow
google-generativeai